Fixat dtos, specifications och mapping

+++ Core/Entities/Orders/OrderItem.cs
@@ -1,10 +1,11 @@
 ﻿using Core.Entities;
+using Core.Entities.Orders;
 
 namespace Core;
 
 public class OrderItem : BaseEntity
 {
     public ItemOrdered ItemOrdered { get; set; } = null!;
-    public double Price { get; set; }
+    public decimal Price { get; set; }
     public int Quantity { get; set; }
 }
